Refactor: Change auth dto hierarchy
