+++ src/main/java/com/petspace/dev/controller/UserController.java
@@ -3,9 +3,9 @@ package com.petspace.dev.controller;
 import com.petspace.dev.dto.user.UserCheckEmailResponseDto;
 import com.petspace.dev.dto.user.UserJoinRequestDto;
 import com.petspace.dev.dto.user.UserLoginRequestDto;
-import com.petspace.dev.dto.user.UserLoginResponseDto;
+import com.petspace.dev.dto.auth.LoginTokenResponseDto;
 import com.petspace.dev.dto.user.UserResponseDto;
-import com.petspace.dev.dto.user.UserTokenReissueRequestDto;
+import com.petspace.dev.dto.auth.LoginTokenReissueRequestDto;
 import com.petspace.dev.service.UserService;
 import com.petspace.dev.util.BaseResponse;
 import lombok.RequiredArgsConstructor;
@@ -40,15 +40,15 @@ public class UserController {
     }
 
     @PostMapping("/login")
-    public BaseResponse<UserLoginResponseDto> login(@Valid @RequestBody UserLoginRequestDto loginRequestDto) {
-        UserLoginResponseDto loginResponseDto = userService.login(loginRequestDto);
+    public BaseResponse<LoginTokenResponseDto> login(@Valid @RequestBody UserLoginRequestDto loginRequestDto) {
+        LoginTokenResponseDto loginResponseDto = userService.login(loginRequestDto);
         return new BaseResponse<>(loginResponseDto);
     }
 
     @PostMapping("/token-reissue")
-    public BaseResponse<UserLoginResponseDto> reissue(@RequestBody UserTokenReissueRequestDto reissueRequestDto) {
+    public BaseResponse<LoginTokenResponseDto> reissue(@RequestBody LoginTokenReissueRequestDto reissueRequestDto) {
         log.info("ref=[{}][{}]", reissueRequestDto.getAccessToken(), reissueRequestDto.getRefreshToken());
-        UserLoginResponseDto loginResponseDto = userService.reissueRefreshToken(reissueRequestDto);
+        LoginTokenResponseDto loginResponseDto = userService.reissueRefreshToken(reissueRequestDto);
         return new BaseResponse<>(loginResponseDto);
     }
 }
