+++ src/main/java/com/petspace/dev/dto/auth/LoginTokenReissueRequestDto.java
@@ -1,9 +1,9 @@
-package com.petspace.dev.dto.user;
+package com.petspace.dev.dto.auth;
 
 import lombok.Getter;
 
 @Getter
-public class UserTokenReissueRequestDto {
+public class LoginTokenReissueRequestDto {
 
     private String accessToken;
     private String refreshToken;
